[Refactor/Python] Refactor python sub-plugin to remove c/cc wrapper
This commit refactors python tensor-filter sub-plugin to remove c/cc
wrapper.
